Eliminated the Spanish language due to difficulties/translation problems..
Yeah, I deleted it because of MANY SEVERE translation problems, and with the technical thing I mean I totally refuse to be 12 hours maybe programming the page again by putting HTML Tags but with the difference that the texts would be of Spanish-speaking language. That is the only true explanation that I have to all this, so I send a greeting to all who read this, and also obviously to those who did not read it xD, Without anything more to say, Goodbye! Cheers - rhssr.